Fix some problem caused by virtual function and XmlRpc::XmlRpcValue

+++ src/logic_filter.cpp
@@ -10,10 +10,10 @@ LogicFilterBase::LogicFilterBase(XmlRpc::XmlRpcValue rpc_value)
 {
   if (rpc_value.hasMember("basic_range") && rpc_value.hasMember("double_check_range"))
   {
-    basic_range_[0] = rpc_value["basic_range"][0];
-    basic_range_[1] = rpc_value["basic_range"][1];
-    double_check_range_[0] = rpc_value["double_check_range"][0];
-    double_check_range_[1] = rpc_value["double_check_range"][1];
+    basic_range_[0] = (double)rpc_value["basic_range"][0];
+    basic_range_[1] = (double)rpc_value["basic_range"][1];
+    double_check_range_[0] = (double)rpc_value["double_check_range"][0];
+    double_check_range_[1] = (double)rpc_value["double_check_range"][1];
   }
   else
     ROS_ERROR("Some filter params doesn't given");
